Add basic structure to webpage in HTML tutorial
A new index.html file has been created for the "Structuring a Web Page" lesson in the HTML Basics section of the HTML/CSS tutorial. The structure includes elements such as header, main, section, aside, and footer, among others. This page will serve as a foundation for the further development of this lesson in the tutorial.