Add various new Python algorithms and update Python SDK
This commit introduces several new Python algorithm files as part of the lesson01 collection. Each file introduces a new algorithm, from guessing a number to calculating crosswalk timing, to Fahrenheit to Celsius conversion and more. In addition, this commit also updates the Python SDK used in the project to Python 3.11. The SDK update provides compatibility with newer Python language features.